Refacoter roomList display function, still need change GUI and function of GUI component
