+++ app/src/main/AndroidManifest.xml
@@ -45,11 +45,6 @@
             android:label="@string/app_name" >
         </activity>
 
-        <activity
-            android:name=".RoomActivity"
-            android:label="@string/app_name" >
-        </activity>
-
     </application>
 
 </manifest>
